Fuction shifted to common place
- Function shifted to common place
